+++ packages/common/src/project/versioned/v0_2_0/types.ts
@@ -1,13 +1,29 @@
 // Copyright 2020-2021 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import {IProjectManifest, SubqlDataSource, SubqlMapping} from '../../types';
+import {
+  SubqlCustomDatasource,
+  SubqlCustomHandler,
+  SubqlDatasource,
+  SubqlDatasourceKind,
+  SubqlHandler,
+  SubqlMapping,
+  SubqlNetworkFilter,
+  SubqlRuntimeDatasource,
+  SubqlRuntimeHandler,
+} from '@subql/types';
+import {IProjectManifest} from '../../types';
 
-export interface SubqlMappingV0_2_0 extends SubqlMapping {
+export interface SubqlMappingV0_2_0<T extends SubqlHandler> extends SubqlMapping<T> {
   file: string;
 }
 
-export type RuntimeDataSourceV0_2_0 = SubqlDataSource<SubqlMappingV0_2_0>;
+export type RuntimeDataSourceV0_2_0 = SubqlRuntimeDatasource<SubqlMappingV0_2_0<SubqlRuntimeHandler>>;
+export type CustomDatasourceV0_2_0 = SubqlCustomDatasource<
+  string,
+  SubqlNetworkFilter,
+  SubqlMappingV0_2_0<SubqlCustomHandler>
+>;
 
 export interface ProjectManifestV0_2_0 extends IProjectManifest {
   name: string;
@@ -24,9 +40,9 @@ export interface ProjectManifestV0_2_0 extends IProjectManifest {
     };
   };
 
-  dataSources: RuntimeDataSourceV0_2_0[];
+  dataSources: (RuntimeDataSourceV0_2_0 | CustomDatasourceV0_2_0)[];
 }
 
-export function isRuntimeDataSourceV0_2_0(dataSource: SubqlDataSource): dataSource is RuntimeDataSourceV0_2_0 {
-  return !!(dataSource as RuntimeDataSourceV0_2_0).mapping.file;
+export function isRuntimeDataSourceV0_2_0(dataSource: SubqlDatasource): dataSource is RuntimeDataSourceV0_2_0 {
+  return dataSource.kind === SubqlDatasourceKind.Runtime && !!(dataSource as RuntimeDataSourceV0_2_0).mapping.file;
 }