+++ packages/common/CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+All logs must start with the format: [x.y.z] - yyyy-mm-dd
+
 ## [Unreleased]
 ### Added
 - Add Bytes type
+
+## [0.12.0] - 2021-10-26
+### Changed
+- Move project manifest types to @subql/types because it is now also required by subquery project. (#512)
 
 ## [0.11.0] - 2021-10-12
 ### Added
@@ -62,7 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - bump dependencies (#148)
 
-## 0.6.0 - 2021-01-27
+## [0.6.0] - 2021-01-27
 ### Fixed
 - pin class-transfermer to 0.3.1 (#116)
 
@@ -79,7 +85,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - support callHandler and eventHandler (#47)
 
-## 0.2.0 - 2020-12-22
+## [0.2.0] - 2020-12-22
 ### Added
 - init commit